Add stripped down version of expr.h and type.h
This commit adds a simplified version of type.h
and expr.h from the previous Relay version.

We implement the basic data types and the
associated machinery for exporting these
to Python, as well as tests that they
can be constructed, all fields are live,
and can be printed using `str`.