Prohibit use mktemp to create temporary files
The temporary file name returned by the mktemp function contains the process ID number, which makes the temporary file name easy to guess. An attacker could exploit this vulnerability to perform a symbolic link attack, allowing a local attacker to overwrite any file, resulting in a denial of service. The alternative function mkdtemp() can be used.