[BugFix][Debug Runtime] fixes debug runtime performance report to correctly handle multi run statistics
Previously all statistics was appended to the list so the result from the first run was always shown while the rest
was just ignored. Found out it because for OpenCL codes first run is usually ignored because of kernel allocation
and compilation that is why it's usually is out of consideration.