Stop pylint complaining about use of function value returning None.
The offending function contains a raise for "not yet supported" but
pylint views the subsequent use of this function as an attempt to use
the value returned from a function returning None.  This patch
temporarily disables the pylint error by adding a dummy return then
disabling the warning on the unreachable return.

This issue could be fixed with a pylint disable markup on each call
site, but this solution in the "not yet implemented" function seems
more appropriate because there is removes the possibility that
dangling pylint markup is left behind at call site when the function
is eventually implemented.