Pin pylint version 2.2.2
pylint is evolving, new version of pylint add new diagnostics.
Pinning the pylint version ensures that the pylint behaviour in a
build environment is consistent between developers and CI and the
environment itself is recreatable.

As pylint evolves the TVM project will need to periodically 'clean'
the code base w.r.t a new current version of pylint and then update
the pinned version.