[CI] Clarify RAT exclude patterns.
RAT exclude patterns are not matched across path separators.  Patterns
of the form xyz/* will match files and directories named xyz anywhere
in the tree because the regexp /* contracts to "".  Adjust the
existing rat exclude patterns that appear to contain matches with path
separators to be explicit about the pattern they actually match.