[CI] RAT exclude patterns for jvm and tutorials
The build process generates artificats into jvm/../targets and
docs/../tutorials subdirectories, many of which trigger RAT.

RAT exclude syntax does not match path separators so it is not
possible to target specific built directories.

This patch tells RAT to ignore and file or directory named 'jvm' and
'tutorials' irrespective of where they are located.

This choice is conservative, and may result in source files being
ingored from the RAT scan.