[VTA] Raise RuntimeError with unknown TARGET
This PR tries to explicitly raise an error when users are trying to compile hardware and forgot to change TARGET  in `vta_config.json`. Specifically, when we try to compile hardware for `de10nano`, and left TARGET in `vta_config.json` to be either `sim` or tsim, previous implementation would quietly return the default configuration for pynq instead of raising an error.