Fixing the bug in loop_partition.
Bounds check can return interval neg_inf, 0 for conditions such as
(1*128 + iter_var*8 + 7) < 130. The bound it finds is
(122 - 128)/6. Here iter_var should not be able to take on value 0 since
with that we are already violating the condition. But since bound check
does div, which by default rounds to zero, we get zero.
This diff fixes this for const int bounds when before division it is
knowns that divisor and dividend are constants and if so it does
floordiv, instead of div. This seems ok for condition on iteration
variables, however not clear if it should do floordiv everywhere else.