[AutoTVM] Added an autotuning Config Library to store autotune results
This implements the first stage of an offline AutoTVM cache,
somewhat like an offline TopHub. Auto-tuning can be performed
in 'jobs' which are a series of sequentially tuned tasks. Data
describing the platform on which the job was performed + when it
took place can also be saved into the library.

This is initial infrastructure to allow for more advanced management
of the auto-tuning process. It includes a mechanism to skip tasks
that are already tuned in the library.

For further information, see the tutorial:
'tutorials/autotvm/tune_with_config_library.py'.