[RELAY] Remove kCompiler attr from ext mod
Functions in external modules kept their kCompiler
attribute which means UseDefaultCompiler returns false.
This prevents you from being able to run any passes on
these external modules. By setting this attribute to
'default' when we add the function to the external
module, we allow for passes to be run external modules.

Change-Id: Ia60de3f0b0db9101738eac407aca4913fad9525a